Trying to make some prototype that generates sound. Have problems with ?distortion?

+++ java/src/andts/javasynth/oscillator/Oscillator.java
@@ -8,9 +8,9 @@ import andts.javasynth.waveform.Waveform;
  */
 public interface Oscillator
 {
-    public double getFrequency();
+    public float getFrequency();
 
-    public void setFrequency(double freq);
+    public void setFrequency(float freq);
 
     public Waveform getWaveform();
